requirements.txt
requests
python-dotenv
beautifulsoup4
tldextract
pandas
